Major update to views.py, models.py, script.js and football.html
Added new model in models.py for class Comment. This is to enable users the chance to comment on posts, changes made to football.html and views.py have led to the user being able to add a comment and submit, the comment goes to the database but unable to view comment at this stage, requires more work.